pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-loop"
version = "0.1.0"
description = "A deterministic, auditable agentic loop built using DSPy"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dspy-ai>=2.5.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
